refactor: rename app-routing to app-routing.module
