+++ src/ng-new/shared/_files/__sourcedir__/app/app.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
-import { AppRoutingModule } from './app-routing';
+import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BarcelonaModule } from './barcelona/barcelona.module';
